+++ src/main/java/com/github/suninvr/virtualadditions/entity/LyftEntity.java
@@ -112,7 +112,7 @@ public class LyftEntity extends AnimalEntity {
     }
 
     @Override
-    public boolean handleFallDamage(float fallDistance, float damageMultiplier, DamageSource damageSource) {
+    public boolean handleFallDamage(double fallDistance, float damageMultiplier, DamageSource damageSource) {
         return false;
     }
